FIX: "save changes" button was getting obscured when previewing themes in preferences
When you preview a theme in preference, some core code is triggered that causes the `padding-bottom` of `#main-outlet` added by the theme to be overriden. To fix this, I'm now using margin instead of padding. This change caused `#topic-progress-wrapper` to be off by the same amount as the margin that was added to `#main-outlet`. That's now fixed by add negative margin to ` #topic-progress-wrapper` to offset for the margin of `#main-outlet`.

TODO: fix Ember 3 and discourse theme settings deprecations once discourse version 2.3 is released.