FEATURE: Add a setting for strict URLs matching
This commit adds a new default-off setting `match_url_params` that, when enabled, makes the component include query params when checking if a page's URL matches a tab. For example, if you have a tab that links to `/latest?order=created`, the component will by default ignore any query params and if you visit `/latest` the tab will be highlighted. However, when `match_url_params` is enabled, you can make the component check for query params as well when determining if a tab should be highlighted.

Thanks to @notedboi for implementing this feature.